error handling added

+++ MyNewApi.Application/Managements/DeleteBannedWord/DeleteBannedWordHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using MyNewApi.Application.Exceptions;
 using MyNewApi.Domain.Interfaces;
 
 namespace MyNewApi.Application.Managements.DeleteBannedWord
@@ -9,6 +10,12 @@ namespace MyNewApi.Application.Managements.DeleteBannedWord
 
         public async Task Handle(DeleteBannedWordRequest request, CancellationToken cancellationToken)
         {
+            var word = await _bannedWordRepository.GetById(request.Word);
+            if (word == null)
+            {
+                throw new NotFoundException($"Word: {request.Word} doesn't exist");
+            }
+
             await _bannedWordRepository.Delete(request.Word);
         }
     }
