+++ MyNewApi.Application/Managements/DeleteProduct/DeleteProductHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using MyNewApi.Application.Exceptions;
 using MyNewApi.Domain.Interfaces;
 
 namespace MyNewApi.Application.Managements.DeleteProduct
@@ -9,6 +10,12 @@ namespace MyNewApi.Application.Managements.DeleteProduct
 
         public async Task Handle(DeleteProductRequest request, CancellationToken cancellationToken)
         {
+            var product = await _productRepository.GetById(request.Id);
+            if (product == null)
+            {
+                throw new NotFoundException($"Product with id: {request.Id} doesn't exist");
+            }
+
             await _productRepository.Delete(request.Id);
         }
     }
